Extracted ImportUriProvider for subclassing in specific domains
